+++ tests/e2e/application-shortcuts.mac.spec.ts
@@ -1,0 +1,7 @@
+import { test } from '@playwright/test'
+import { applicationShortcutTests } from './application-shortcuts.helpers'
+
+test.describe('macOS application shortcuts', () => {
+  test.skip(process.platform !== 'darwin', 'macOS only')
+  applicationShortcutTests()
+})